requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
openai==1.60.0
requests==2.32.0
python-dotenv==1.0.1
jinja2==3.1.4
pytest==8.3.0
